# Performance optimizations to reduce large repository loading time