@hjuutilainen 2025-10-02
Thank you for this, looks interesting and the performance improvements are significant! I wanted to get the current release out before taking a closer look at this. Will get back to you soon.